Изменения по проекту на 31.12.2019
