+++ src/components/abstract-component/abstract-component.js
@@ -1,4 +1,4 @@
-import {createElement} from '../../utils';
+import {createElement} from '../../utils/render';
 
 export default class AbstractComponent {
   constructor() {
