+++ src/components/task-edit/task-edit.js
@@ -1,5 +1,5 @@
 import {Colors, MonthNames, Days} from '../../const';
-import {formatTime} from '../../utils.js';
+import {formatTime} from '../../utils/common';
 import AbstractComponent from '../abstract-component/abstract-component';
 
 const createTagTemplate = (tag) => {
